Updaed domain command to have long ids for word, details, relations and translations #3

+++ src/Domain/Commands/AddWordDetailCommand.cs
@@ -4,7 +4,7 @@ namespace Inshapardaz.Domain.Commands
 {
     public class AddWordDetailCommand : Command
     {
-        public int WordId { get; set; }
+        public long WordId { get; set; }
 
         public WordDetail WordDetail { get; set; }
     }
