+++ src/Domain/Commands/AddWordMeaningCommand.cs
@@ -3,9 +3,9 @@ using Inshapardaz.Domain.Model;
 
 namespace Inshapardaz.Domain.Commands
 {
-    public class AddWordMeaningCommand  : Command
+    public class AddWordMeaningCommand : Command
     {
-        public int WordDetailId { get; set; }
+        public long WordDetailId { get; set; }
 
         public Meaning Meaning { get; set; }
     }
