+++ src/Domain/Commands/AddWordRelationCommand.cs
@@ -4,8 +4,8 @@ namespace Inshapardaz.Domain.Commands
 {
     public class AddWordRelationCommand : Command
     {
-        public int SourceWordId { get; set; }
-        public int RelatedWordId { get; set; }
+        public long SourceWordId { get; set; }
+        public long RelatedWordId { get; set; }
         public RelationType RelationType { get; set; }
     }
 }
