+++ src/Domain/Commands/AddWordTranslationCommand.cs
@@ -4,9 +4,9 @@ namespace Inshapardaz.Domain.Commands
 {
     public class AddWordTranslationCommand : Command
     {
-        public int WordId { get; set; }
+        public long WordId { get; set; }
 
-        public int WordDetailId  { get; set; }
+        public long WordDetailId { get; set; }
 
         public Translation Translation { get; set; }
     }
