+++ src/Domain/Queries/TranslationsByLanguageQuery.cs
@@ -6,7 +6,7 @@ namespace Inshapardaz.Domain.Queries
 {
     public class TranslationsByLanguageQuery : IQuery<IEnumerable<Translation>>
     {
-        public int WordId { get; set; }
+        public long WordId { get; set; }
 
         public Languages Language { get; set; }
     }
